Moved ExercisePanel picker code to picker.js
